requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.1.7
django-crispy-forms==2.0
django-tinymce==3.5.0
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
